requirements.txt
simplejson
lxml
codecov
coverage
flask==2.0.0
pre - commit
pytest
PyJWT
requests
